+++ LevelGeneration/Assets/Features/ProceduralTerrainGeneration/Scenes/TerrainGeneration.unity
@@ -366,7 +366,7 @@ MeshFilter:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1113364076}
-  m_Mesh: {fileID: 1530139747}
+  m_Mesh: {fileID: 1599050746}
 --- !u!1 &1379970720
 GameObject:
   m_ObjectHideFlags: 0
@@ -444,7 +444,7 @@ Transform:
   m_GameObject: {fileID: 1379970720}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 12.5, y: 0.1, z: 12.5}
+  m_LocalScale: {x: 5.3, y: 0.1, z: 5.3}
   m_Children: []
   m_Father: {fileID: 758110413}
   m_RootOrder: 0
@@ -461,7 +461,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 5d2c0bbda3540ab4f878256d2638b7f8, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!43 &1530139747
+--- !u!43 &1599050746
 Mesh:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
