+++ LevelGeneration/Assets/Scripts/Utility/HideOnPlay.cs
